Fix For Issue 6 (With Test)
Main has the testing on this commit, note I was testing with ../sample/ as the directories are setup, not ../testing/ as was stated in issue, but it's the same effect.

The changes in Graph.cpp update std::filesystem::recursive_directory_iterator to convert paths based on the dirPath provided (now supplied as an absolute path named root). Instead of the default behavior which is where the executable is.